feat: add mainLayout
