+++ src/layouts/mainLayout/mainLayout.module.scss
@@ -1,0 +1,22 @@
+.wrapper {
+  background-image: var(--background-main-page-image);
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center;
+  width: 100%;
+}
+
+.main_layout {
+  width: 100%;
+  min-height: 100vh;
+  background-color: inherit;
+  max-width: 1440px;
+  margin: 0 auto;
+
+  .main {
+      background: red;
+      display: flex;
+      flex-direction: column;
+      min-height: 100vh;
+  }
+}
