--- src/pages/main-page/components/contentComponent/index.ts
@@ -1,1 +1,0 @@
-export { ContentComponent } from './contentComponent';
