+++ src/pages/mainPage/components/contentComponent/components/actionsCardComponent/actionCardComponent.module.scss
