+++ src/pages/mainPage/components/contentComponent/components/actionsCardComponent/actionsCardComponent.tsx
