+++ src/pages/mainPage/components/contentComponent/components/textCardComponent/index.ts
