+++ src/pages/mainPage/components/contentComponent/components/textCardComponent/textCardComponent.tsx
