+++ src/pages/mainPage/components/contentComponent/components/textCardComponent/textCardContent.module.scss
