+++ src/pages/mainPage/components/contentComponent/contentComponent.tsx
