+++ src/pages/mainPage/components/contentComponent/index.ts
@@ -1,0 +1,1 @@
+export { ContentComponent } from './contentComponent';
