+++ src/pages/mainPage/components/footerComponent/components/footerCardComponent.module.scss
