+++ src/pages/mainPage/components/footerComponent/components/index.ts
