+++ src/pages/mainPage/components/footerComponent/index.ts
