+++ src/pages/mainPage/components/headerComponent/headerComponent.module.scss
