+++ src/pages/mainPage/components/headerComponent/headerComponent.tsx
