+++ src/pages/mainPage/components/headerComponent/index.ts
