+++ src/pages/mainPage/components/siderComponent/index.ts
