+++ src/pages/mainPage/components/siderComponent/siderComponent.module.scss
