+++ src/pages/mainPage/components/siderComponent/siderComponent.tsx
