+++ src/pages/mainPage/index.ts
